requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pandas
pydantic
python-multipart
psycopg2-binary       # ← PostgreSQL 드라이버
python-dotenv         # ← .env 사용 시
scikit-learn
openai
requests